Use `weak!` macro instead of `clone!` to pass the `ApplicationController` into callback closures
Previously, the `Rc<...>` for the `ApplicationController` was cloned and
the clone moved into the closures for the Gtk callbacks. This caused a
`ApplicationController` not to be dropped prior to program termination
because the closures still kept a strong shared pointer on it.
(Possibly this was even a cyclical dependency but I'm not quite sure of
the ownership model of Gtk's callbacks.) To avoid this, a `Weak<...>`
pointer is now moved into the closures instead, so as to avoid them
blocking destruction of the `ApplicationController`. This is
facilitated by a new macro, `weak!`.