Add support for iteration parameters
This has some caveats still. For one, iteration parameters are always
passed as strings. This works fine with MERLIC which does not enforce
matching types as long as they are losslessly convertible. It would be
better though to pass them in the specified recipe parameter type
(and to validate input accordingly).